Improve the RESTful API for User entity

+++ src/main/java/com/pcs/poseidon/controllers/UserController.java
@@ -15,6 +15,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import jakarta.validation.Valid;
 
+/**
+ * The CRUD controller handling HTTP requests for managing users.
+ */
 @Controller
 @RequiredArgsConstructor
 @Slf4j
@@ -46,7 +49,7 @@ public class UserController {
     }
 
     @GetMapping("/users/update/{id}")
-    public String showUpdateForm(@PathVariable("id") Integer id, Model model) {
+    public String showUpdateForm(@PathVariable("id") Long id, Model model) {
         var user = userRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
         user.setPassword("");
         model.addAttribute("user", user);
@@ -54,7 +57,7 @@ public class UserController {
     }
 
     @PostMapping("/users/update/{id}")
-    public String updateUser(@PathVariable("id") Integer id, @Valid User user, BindingResult result, Model model) {
+    public String updateUser(@PathVariable("id") Long id, @Valid User user, BindingResult result, Model model) {
         if (result.hasErrors()) {
             return "users/update";
         }
@@ -68,7 +71,7 @@ public class UserController {
     }
 
     @GetMapping("/users/delete/{id}")
-    public String deleteUser(@PathVariable("id") Integer id, Model model) {
+    public String deleteUser(@PathVariable("id") Long id, Model model) {
         var user = userRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
         userRepository.delete(user);
         model.addAttribute("users", userRepository.findAll());
